Rename Resource -> Interaction, because:
Our typical use case for a test case (= scenario) will consist of
a series of interactions, where each interaction is defined as a
request and an expected response.

In this use case we could very well make requests to the same resource
but using e.g. different request methods in the different steps
(= interactions), and thereby expecting different responses in each
interaction.